Add interactive console utility
